% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Calendar Visualization}
\usage{
calendar(
  data = NULL,
  from = NULL,
  to = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Chart data}

\item{from}{start date}

\item{to}{end date}

\item{...}{additional arguments}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{

}
\description{
Calendar Visualization
}
\examples{
library(nivor)

df <- data.frame(
  day = seq.Date(
    from = as.Date("2017-03-15"),
    length.out = 500,
    by = "days"
  ),
  value = round(runif(500) * 1000, 0)
)

calendar(df)
}
\seealso{
\href{https://nivo.rocks/calendar/}{Additional arguments}
}
