% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{<Add Title>}
\usage{
choropleth(
  data,
  polygon_json,
  projectionType = c("azimuthalEqualArea", "azimuthalEquidistant", "gnomonic",
    "orthographic", "stereographic", "equalEarth", "equirectangular", "mercator",
    "transverseMercator", "naturalEarth1"),
  projectionScale = 200,
  interective_rotate = FALSE,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\description{
<Add Description>
}
