% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bump.R
\name{n_bump}
\alias{n_bump}
\title{Bump Visualization}
\usage{
n_bump(
  data = NULL,
  series = NULL,
  x = NULL,
  ...,
  width = NULL,
  height = NULL,
  element_id = NULL
)
}
\arguments{
\item{data}{data set in data.frame.}

\item{series, x}{column names of series and x value. if not provided, it will
use the first column as x value, and others as series.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{element_id}{element id of widgets.}
}
\value{
a nivo Bump component
}
\description{
The Bump chart can be used to show the ranking of several series over time.
It is quite similar to line charts, but instead of graphing some measure on
the y-axis, it only shows the ranking of each serie at a given time.
}
\details{
If you'd like to show the ranking and also graph the y-axis values, you can
also you use the AreaBump visualization.
}
\seealso{
\href{https://nivo.rocks/bump/}{Additional arguments}
}
