% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chord.R
\name{n_chord}
\alias{n_chord}
\title{Chord Visualization}
\usage{
n_chord(
  matrix,
  keys,
  render = c("svg", "canvas"),
  ...,
  width = NULL,
  height = NULL,
  element_id = NULL
)
}
\arguments{
\item{matrix}{a matrix used to compute the chord diagram.}

\item{keys}{keys used to identify each cell in the matrix.}

\item{render}{"svg" (responsive) or "canvas". "canvas" is well suited for
large data sets as it does not impact DOM tree depth, however you'll lose the
isomorphic rendering ability.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{element_id}{element id of widgets.}
}
\value{
a nivo Chord component
}
\description{
Chord diagram.
}
\examples{
library(nivor)

# generate data
set.seed(1)
data <- matrix(round(rexp(25) * 100, 0), 5, 5)

# the simplest use
n_chord(
  matrix = data,
  keys = c("A", "B", "C", "D", "E")
)
}
\seealso{
\href{https://nivo.rocks/chord/}{Additional arguments}
}
