% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{n_options}
\alias{n_options}
\title{Add additional settings to the default}
\usage{
n_options(default, ...)
}
\arguments{
\item{default}{some common or default setting}

\item{...}{additional options for nivo components}
}
\value{
a list contains settings for nivo components
}
\description{
Add additional settings to the default
}
