% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{n_parallel_coordinates}
\alias{n_parallel_coordinates}
\title{Parallel coordinates chart}
\usage{
n_parallel_coordinates(
  data,
  variables,
  render = c("svg", "canvas"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a data.frame contains value for create the pie}

\item{variables}{variables configuration.}

\item{render}{"svg" (responsive) or "canvas". "canvas" is well suited
for large data sets as it does not impact DOM tree depth, however you'll lose
the isomorphic rendering ability.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{
a parallel coordinates component
}
\description{
Parallel coordinates chart, supports continuous numerical variables and
discrete values.
}
\examples{
library(nivor)

# the simplest usage
n_parallel_coordinates(data = iris)

# generate data
set.seed(1)
data <- data.frame(
  temp = round(rexp(32) * 10, 0),
  cost = round(rexp(32) * 10000, 0),
  color = sample(c("red", "yellow", "green"), size = 32, replace = TRUE),
  target = sample(c("A", "B", "C", "D", "E"), size = 32, replace = TRUE),
  volume = round(rexp(32) * 3, 0)
)

# use with default options
n_parallel_coordinates(data = data)

# customized, only show four series
n_parallel_coordinates(
  data = data,
  variables = list(
    list(
      key = "temp", type = "linear",
      legend = "Temperature", legendPosition = "start", legendOffset = -20
    ),
    list(
      key = "cost", type = "linear", legend = "Cost", min = 0,
      legendOffset = -20
    ),
    list(key = "color", type = "point", padding = 1),
    list(key = "target", type = "point", values = c("A", "B", "D", "E"))
  )
)
}
\seealso{
\href{https://nivo.rocks/parallel-coordinates/}{Additional
arguments}
}
