% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream.R
\name{n_stream}
\alias{n_stream}
\title{Stream chart}
\usage{
n_stream(data, keys, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data.frame contains value for create the stream chart}

\item{keys}{column name for grouping.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{
a sStream component
}
\description{
Stream chart
}
\examples{
library(nivor)

# the simplest usage
n_parallel_coordinates(data = iris)

# generate data
set.seed(1)
data <- data.frame(
  temp = round(rexp(32) * 10, 0),
  cost = round(rexp(32) * 10000, 0),
  color = sample(c("red", "yellow", "green"), size = 32, replace = TRUE),
  target = sample(c("A", "B", "C", "D", "E"), size = 32, replace = TRUE),
  volume = round(rexp(32) * 3, 0)
)

# use with default options
n_parallel_coordinates(data = data)

# customized, only show four series
n_parallel_coordinates(
  data = data,
  variables = list(
    list(
      key = "temp", type = "linear",
      legend = "Temperature", legendPosition = "start", legendOffset = -20
    ),
    list(
      key = "cost", type = "linear", legend = "Cost", min = 0,
      legendOffset = -20
    ),
    list(key = "color", type = "point", padding = 1),
    list(key = "target", type = "point", values = c("A", "B", "D", "E"))
  )
)
}
\seealso{
\href{https://nivo.rocks/stream/}{Additional
arguments}
}
