% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/voronoi.R
\name{n_voronoi}
\alias{n_voronoi}
\title{Voronoi Visualization}
\usage{
n_voronoi(data, id, x, y, ..., width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{a data.frame contains value for create the voronoi.}

\item{id}{column name of id.}

\item{x, y}{column names of x and y value.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{
a nivo Voronoi component
}
\description{
Delaunay/Voronoi Tessellation.
}
\examples{
library(nivor)

# generate data
data <- data.frame(
  id = 1:50,
  x = round(rexp(50) * 20, 0),
  y = round(rexp(50) * 20, 0)
)

# the simplest use
n_voronoi(
  data = data,
  id = "id",
  x = "x",
  y = "y",
  xDomain = c(0, 100),
  yDomain = c(0, 100),
  enableLinks = TRUE,
  linkLineColor = "#cccccc",
  cellLineColor = "#c6432d",
  pointSize = 6,
  pointColor = "#c6432d"
)

}
\seealso{
\href{https://nivo.rocks/voronoi/}{Additional arguments}
}
