% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle.R
\name{n_waffle}
\alias{n_waffle}
\title{Waffle Visualization}
\usage{
n_waffle(
  data,
  id = NULL,
  value = NULL,
  label = NULL,
  total,
  rows,
  columns,
  render = c("svg", "html", "canvas"),
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a data.frame contains value for create the waffle}

\item{id}{the name of the column used to represent the group.}

\item{value}{column names for value.}

\item{label}{column names for label. If not provided, id is used as the label
by default.}

\item{total}{max value.}

\item{rows}{number of rows.}

\item{columns}{number of columns.}

\item{render}{"svg" (responsive) "html", or "canvas". "canvas" is well suited
for large data sets as it does not impact DOM tree depth, however you'll lose
the isomorphic rendering ability.}

\item{...}{additional arguments.}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{
a waffle calendar component
}
\description{
Waffle Visualization
}
\examples{
library(nivor)

# generate data
data <- data.frame(
  id = c("men", "women", "children"),
  label = c("Men", "Women", "Children"),
  value = c(9, 13, 20),
  color = c("#468df3", "#ba72ff", "#a1cfff")
)

# the simplest use
n_waffle(
  data = data,
  total = 100,
  rows = 18,
  columns = 14
)

# render with customization.
n_waffle(
  data = data,
  total = 100,
  rows = 18,
  columns = 14,
  margin = list(
    top = 10,
    right = 10,
    bottom = 10,
    left = 120
  ),
  colors = list(scheme = "nivo"),
  borderColor = list(
    from = "color",
    modifiers = list(
      c("darker", 1.3)
    )
  ),
  animate = TRUE,
  motionStiffness = 90,
  motionDamping = 11,
  legends = list(
    list(
      anchor = "top-left",
      direction = "column",
      justify = FALSE,
      translateX = -100,
      translateY = 0,
      itemsSpacing = 4,
      itemWidth = 100,
      itemHeight = 20,
      itemDirection = "left-to-right",
      itemOpacity = 1,
      itemTextColor = "#777",
      symbolSize = 20,
      effects = list(
        list(
          on = "hover",
          style = list(
            itemTextColor = "#000",
            itemBackground = "#f7fafb"
          )
        )
      )
    )
  )
)
}
\seealso{
\href{https://nivo.rocks/waffle/}{Additional arguments}
}
