% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nivocal.R
\name{nivocal}
\alias{nivocal}
\title{Calendar Visualization}
\usage{
nivocal(
  data = NULL,
  from = NULL,
  to = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{Chart data}

\item{from}{start date}

\item{to}{end date}

\item{...}{additional arguments}

\item{width}{width of widgets}

\item{height}{height of widgets}

\item{elementId}{element id of widgets}
}
\value{

}
\description{
Calendar Visualization
}
\examples{
library(nivor)

df <- data.frame(
  day = seq.Date(
    from = as.Date("2017-03-15"),
    length.out = 500,
    by = "days"
  ),
  value = round(runif(500) * 1000, 0)
)

nivocal(df)
}
\seealso{
\href{https://nivo.rocks/calendar/}{Additional arguments}
}
