% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Calendar Visualization}
\usage{
calendar(
  data = NULL,
  day = NULL,
  value = NULL,
  from = NULL,
  to = NULL,
  ...,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{a data.frame contains day and value for create the calendar.}

\item{day, value}{column names of day and value. if not provided, it will use
the first column as day, and the second column as value.}

\item{from}{start date}

\item{to}{end date}

\item{...}{additional arguments}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{elementId}{element id of widgets}
}
\value{
a nivo calendar component
}
\description{
Calendar Visualization
}
\examples{
library(nivor)

df <- data.frame(
  day = seq.Date(
    from = as.Date("2020-04-01"),
    length.out = 600,
    by = "days"
  ),
  value = round(runif(600) * 1000, 0)
)

# default
calendar(df)

# GitHub style
calendar(
  data = df,
  emptyColor = "#eeeeee",
  colors = c("#d6e685", "#8cc665", "#44a340", "#1e6823"),
  yearSpacing = 40,
  monthBorderColor = "#ffffff",
  dayBorderWidth = 2,
  dayBorderColor = "#ffffff"
)
}
\seealso{
\href{https://nivo.rocks/calendar/}{Additional arguments}
}
