% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choropleth.R
\name{choropleth}
\alias{choropleth}
\title{Choropleth}
\usage{
choropleth(
  map_id,
  data,
  polygon_json = nivoR::world_countries_json,
  projectionType = c("azimuthalEqualArea", "azimuthalEquidistant", "gnomonic",
    "orthographic", "stereographic", "equalEarth", "equirectangular", "mercator",
    "transverseMercator", "naturalEarth1"),
  projectionScale = 200,
  projectionTranslation_x = 0.5,
  projectionTranslation_y = 0.5,
  rotate_x = 0,
  rotate_y = 0,
  rotate_z = 0,
  domainMin = 0,
  domainMax = 1e+06,
  border_width = 0.5,
  border_Color = "#989627",
  width = NULL,
  height = "400px",
  margin_top = 0,
  margin_left = 0,
  margin_right = 0,
  margin_bottom = 0,
  unknownColor = "#666666",
  tooltipText,
  polygonColors = "#9b0034",
  pausePlayBTNcolor = "black",
  elementId = NULL,
  ...
)
}
\description{
\if{html}{\out{
<Add Description>
}}
}
